qute: Command with completion to manipulate rules.
This adds a command for qutebrowser to toggle block rules. Technically
you can add new rules via it as well since the command doesn't rely on
the completion model and just uses the entered text.

The completion is scoped to the host of the current tab, although it
might look like it is scoped to the tab itself.

The completion is filled from a new Rules object that tracks all seen
requests (actually all soon origin-destination-types so it isn't quite
that bad). It might be nice to clean it up in some way but I don't see a
way to do so without getting a signal when any tab closes or changes URL
and then going over open tabs and seeing if any of the origins aren't in
use anymore.